お題変更反映 リレーション追加

+++ app/Http/Controllers/Admin/ThemeController.php
@@ -20,9 +20,30 @@ class ThemeController extends Controller
 
     public function index() {
 
-        $themes = Theme::all()->sortByDesc('created_at');
+        $themes = Theme::all();
+        $applyTheme  = Theme::where('apply', 1)->first();
 
-        return view('admin.theme.index', ['themes' => $themes]);
+        return view('admin.theme.index', [
+            'themes' => $themes,
+            'applyTheme'  => $applyTheme,
+        ]);
+    }
+
+    public function edit(Theme $theme)
+    {        
+        return view('admin.theme.edit', ['theme' => $theme,]);
+    }
+
+    public function update(Theme $theme)
+    {
+        $applyTheme  = Theme::where('apply', 1)->first();
+        
+        $applyTheme->update(['apply' => 0]);
+        $theme->update(['apply' => 1]);
+
+        $theme->save();
+
+        return redirect()->route('admin.themes.index');
     }
 
     public function create() {
@@ -50,6 +71,9 @@ class ThemeController extends Controller
 
     public function destroy(Theme $theme) {
 
+        if($theme->apply === 1){
+            abort(404, '消せません');
+        }
         $theme->delete();
 
         return redirect()->route('admin.themes.index');
